improvement: added trace spans for asynchronous workflow executions, previously only appeared when workflow was executed manually
